#563 JR East single color line style
